Add support and 3 customizations for touch based devices.
This game is only for desktop machines. We need to handle touch
events in JavaScript to make it work on touch based mobile devices.
Add listeners for touch events and process them to control the
directions of the vehicle.

Besides, it provides control hints on screen because it's not obvious
for touch based screen to how to control the vehicle. And remove
the score board and set the game quality as a low level due to
the limitation of mobile devices.

Contributions here are licensed under 'Creative Commons Attribution
-NonCommercial 3.0 Unported License'.

Signed-off-by: Harry Shan <zhenyu.shan@intel.com>